Prevent reloading of settings page at saving or syncing.
Some browsers like Firefox do not wait for event handler callbacks to
finish when a navigation event has been propagated.
This behavior prevents the execution of the settings sync, which
executes the sync after saving.
To mitigate this the type of the HTML button elements are set to
'button' (default: 'submit') which does not propagate a navigation
event.